requirements.txt
alembic==1.0.5
amqp==2.4.0
APScheduler==3.5.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.5
Deprecated==1.2.4
factory-boy==2.11.1
Faker==1.0.1
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
passlib==1.7.1
pluggy==0.8.0
py==1.7.0
pycparser==2.19
PyGithub==1.43.4
PyJWT==1.7.1
PyMySQL==0.9.3
pytest==4.0.2
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.7
requests==2.21.0
six==1.12.0
soupsieve==1.7
speaklater==1.3
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.10
text-unidecode==1.2
tzlocal==1.5.1
urllib3==1.24.1
vine==1.2.0
waitress==1.1.0
WebOb==1.8.5
WebTest==2.0.32
Werkzeug==0.14.1
wrapt==1.11.0
WTForms==2.2.1
